+++ src/pybricks/robotics.py
@@ -8,7 +8,7 @@ from __future__ import annotations
 from typing import Tuple, Optional, overload, TYPE_CHECKING
 
 from . import _common
-from .parameters import Stop
+from .parameters import Stop, Direction
 
 if TYPE_CHECKING:
     from ._common import Motor
@@ -29,8 +29,7 @@ class DriveBase:
     **Positive** angles and turn rates mean turning **right**.
     **Negative** means **left**. So when viewed from the top,
     positive means clockwise and negative means counterclockwise. If desired,
-    you can flip this convention by reversing the ``left_motor`` and
-    ``right_motor`` when you initialize this class.
+    you can reverse this behavior with the ``positive_direction`` parameter.
 
     See the `measuring`_ section for tips to measure and adjust the diameter
     and axle track values.
@@ -59,8 +58,9 @@ class DriveBase:
         right_motor: Motor,
         wheel_diameter: Number,
         axle_track: Number,
+        positive_direction: Direction = Direction.CLOCKWISE,
     ):
-        """DriveBase(left_motor, right_motor, wheel_diameter, axle_track)
+        """DriveBase(left_motor, right_motor, wheel_diameter, axle_track, positive_direction=Direction.CLOCKWISE)
 
         Arguments:
             left_motor (Motor):
@@ -70,6 +70,9 @@ class DriveBase:
             wheel_diameter (Number, mm): Diameter of the wheels.
             axle_track (Number, mm): Distance between the points where
                 both wheels touch the ground.
+            positive_direction (Direction): Which direction the drive base
+                should turn when you give a positive turn rate or turn
+                angle, viewed from the top.
         """
 
     def drive(self, speed: Number, turn_rate: Number) -> None:
@@ -135,25 +138,30 @@ class DriveBase:
         ...
 
     def settings(self, *args):
-        """settings(straight_speed, straight_acceleration, turn_rate, turn_acceleration)
+        """
+        settings(straight_speed, straight_acceleration, turn_rate, turn_acceleration)
         settings() -> Tuple[int, int, int, int]
 
-        Configures the speed and acceleration used
-        by :meth:`.straight`, :meth:`.turn`, and  :meth:`.curve`.
+        Configures the drive base speed and acceleration.
 
         If you give no arguments, this returns the current values as a tuple.
 
-        The default values are automatically configured based on your wheel
+        The initial values are automatically configured based on your wheel
         diameter and axle track. They are selected such that your robot
         drives at about 40% of its maximum speed.
+
+        The speed values given here do not apply to the :meth:`.drive` method,
+        since you provide your own speed values as arguments in that method.
 
         Arguments:
             straight_speed (Number, mm/s): Straight-line speed of the robot.
             straight_acceleration (Number, mm/s²): Straight-line
-                acceleration and deceleration of the robot.
+                acceleration and deceleration of the robot. Provide a tuple with
+                two values to set acceleration and deceleration separately.
             turn_rate (Number, deg/s): Turn rate of the robot.
             turn_acceleration (Number, deg/s²): Angular acceleration and
-                deceleration of the robot.
+                deceleration of the robot. Provide a tuple with
+                two values to set acceleration and deceleration separately.
         """
 
     def straight(
@@ -221,6 +229,7 @@ class DriveBase:
 
 # HACK: hide from jedi
 if TYPE_CHECKING:
+    del Direction
     del Motor
     del Number
     del Stop